+++ src/edit.js
@@ -7,8 +7,10 @@ var MAX_URL_LENGTH = 2000
 var cachedBlock = null
 var textarea = document.getElementById('textarea')
 var status = document.getElementById('status')
+var heightInput = document.getElementById('height')
 var cachedHeight = 0
 var hideStatusTimeout = null
+var height = 240
 
 function showStatus (text) {
   if (hideStatusTimeout) {
@@ -47,8 +49,11 @@ function toEdit (block) {
   }
   block.metadata.widget = 'userhtml'
   cachedBlock = block
-  if (block.hasOwnProperty('text')) {
-    textarea.value = block.text
+  if (block.text || block.metadata.text) {
+    textarea.value = block.text || block.metadata.text
+  }
+  if (block.height || block.metadata.height) {
+    heightInput.value = block.height || block.metadata.height
   }
   resize()
 }
@@ -58,17 +63,23 @@ function fromEdit (event) {
 
   showStatus('saving...')
 
-  var text = event.target.value
+  var text = textarea.value
+  var height = parseInt(heightInput.value, 10)
+  height = Math.min(5000, Math.max(1, height))
   var b64 = b64ify(text)
 
   cachedBlock.text = text
+  cachedBlock.metadata.text = text
+  cachedBlock.height = height
+  cachedBlock.metadata.height = height
+
   var url = RUNNER + b64
   if (url.length > MAX_URL_LENGTH) {
     showStatus('error: input too long')
     return
   }
   cachedBlock.metadata.isBasedOnUrl = url
-  cachedBlock.html = '<iframe src="'+url+'"></iframe>'
+  cachedBlock.html = '<iframe src="'+url+'" height="'+height+'"></iframe>'
 
   hideStatusTimeout = setTimeout(hideStatus, 1000)
 
@@ -76,12 +87,13 @@ function fromEdit (event) {
   resize()
 }
 textarea.addEventListener('input', fromEdit)
+heightInput.addEventListener('input', fromEdit)
 
 function resize () {
   textarea.style.height = 'auto'
   var height = textarea.scrollHeight
-  // Padding is 2px, so minus 4
-  textarea.style.height = (height - 4) + 'px'
+  textarea.style.height = height + 'px'
+  height += 30
   if (height !== cachedHeight) {
     cachedHeight = height
     send('height', height)