+++ CHANGES.md
@@ -1,3 +1,9 @@
+## 0.2.0, 2016-08-23
+
+* Read and write both `block.metadata.text` and `block.text`
+* Add UI to set `height`: saves to html, block, and metadata
+* Support link
+
 ## 0.1.0, 2016-08-23
 
 * Saves text to `block.text` (instead of `metadata`)